feat: memperbarui halaman node viewer
