+++ server/views/runner/htmlViewer.ejs
@@ -129,6 +129,7 @@
     <script src="/assets/prismjs/plugins/autoloader/prism-autoloader.min.js"></script>
     <script src="/assets/prismjs/plugins/line-numbers/prism-line-numbers.min.js"></script>
 
+    <script src="/public/js/common.js"></script>
     <script src="/public/js/htmlViewer.js"></script>
   </body>
 </html>
